Extract radio logic to subcomponent
